refactor: compacting things
